move BasicEvaluatedExpression and JavascriptParserHelpers to javascript directory
