+++ lib/APIPlugin.js
@@ -5,13 +5,13 @@
 
 "use strict";
 
-const {
-	toConstantDependency,
-	evaluateToString
-} = require("./JavascriptParserHelpers");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const WebpackError = require("./WebpackError");
 const ConstDependency = require("./dependencies/ConstDependency");
+const {
+	toConstantDependency,
+	evaluateToString
+} = require("./javascript/JavascriptParserHelpers");
 const ChunkNameRuntimeModule = require("./runtime/ChunkNameRuntimeModule");
 const GetFullHashRuntimeModule = require("./runtime/GetFullHashRuntimeModule");
