+++ lib/CommonJsStuffPlugin.js
@@ -5,16 +5,16 @@
 
 "use strict";
 
-const {
-	evaluateToIdentifier,
-	expressionIsUnsupported,
-	toConstantDependency
-} = require("./JavascriptParserHelpers");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const RuntimeModule = require("./RuntimeModule");
 const Template = require("./Template");
 const ModuleDecoratorDependency = require("./dependencies/ModuleDecoratorDependency");
 const RuntimeRequirementsDependency = require("./dependencies/RuntimeRequirementsDependency");
+const {
+	evaluateToIdentifier,
+	expressionIsUnsupported,
+	toConstantDependency
+} = require("./javascript/JavascriptParserHelpers");
 
 class CommonJsStuffPlugin {
 	apply(compiler) {
