+++ lib/ConstPlugin.js
@@ -5,9 +5,9 @@
 
 "use strict";
 
-const { evaluateToString } = require("./JavascriptParserHelpers");
 const CachedConstDependency = require("./dependencies/CachedConstDependency");
 const ConstDependency = require("./dependencies/ConstDependency");
+const { evaluateToString } = require("./javascript/JavascriptParserHelpers");
 
 /** @typedef {import("./Compiler")} Compiler */
