+++ lib/DefinePlugin.js
@@ -5,14 +5,14 @@
 
 "use strict";
 
-const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
+const RuntimeGlobals = require("./RuntimeGlobals");
+const ConstDependency = require("./dependencies/ConstDependency");
+const BasicEvaluatedExpression = require("./javascript/BasicEvaluatedExpression");
 const {
 	approve,
 	evaluateToString,
 	toConstantDependency
-} = require("./JavascriptParserHelpers");
-const RuntimeGlobals = require("./RuntimeGlobals");
-const ConstDependency = require("./dependencies/ConstDependency");
+} = require("./javascript/JavascriptParserHelpers");
 
 /** @typedef {import("./Compiler")} Compiler */
 /** @typedef {import("./javascript/JavascriptParser")} JavascriptParser */
