+++ lib/HotModuleReplacementPlugin.js
@@ -8,11 +8,6 @@
 const { SyncBailHook } = require("tapable");
 const { RawSource } = require("webpack-sources");
 const HotUpdateChunk = require("./HotUpdateChunk");
-const {
-	evaluateToIdentifier,
-	evaluateToString,
-	toConstantDependency
-} = require("./JavascriptParserHelpers");
 const NormalModule = require("./NormalModule");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const ModuleHotAcceptDependency = require("./dependencies/ModuleHotAcceptDependency");
@@ -20,6 +15,11 @@ const ModuleHotDeclineDependency = require("./dependencies/ModuleHotDeclineDepen
 const ModuleHotDependency = require("./dependencies/ModuleHotDependency");
 const HotModuleReplacementRuntimeModule = require("./hmr/HotModuleReplacementRuntimeModule");
 const JavascriptParser = require("./javascript/JavascriptParser");
+const {
+	evaluateToIdentifier,
+	evaluateToString,
+	toConstantDependency
+} = require("./javascript/JavascriptParserHelpers");
 const { find } = require("./util/SetHelpers");
 const { compareModulesById } = require("./util/comparators");
