+++ lib/NodeStuffPlugin.js
@@ -5,11 +5,11 @@
 
 "use strict";
 
+const CachedConstDependency = require("./dependencies/CachedConstDependency");
 const {
 	evaluateToString,
 	expressionIsUnsupported
-} = require("./JavascriptParserHelpers");
-const CachedConstDependency = require("./dependencies/CachedConstDependency");
+} = require("./javascript/JavascriptParserHelpers");
 const { relative } = require("./util/fs");
 
 /** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
