+++ lib/ProvidePlugin.js
@@ -5,9 +5,9 @@
 
 "use strict";
 
-const { approve } = require("./JavascriptParserHelpers");
 const ConstDependency = require("./dependencies/ConstDependency");
 const ProvidedDependency = require("./dependencies/ProvidedDependency");
+const { approve } = require("./javascript/JavascriptParserHelpers");
 
 /** @typedef {import("./Compiler")} Compiler */
