+++ lib/RequireJsStuffPlugin.js
@@ -5,9 +5,11 @@
 
 "use strict";
 
-const { toConstantDependency } = require("./JavascriptParserHelpers");
 const RuntimeGlobals = require("./RuntimeGlobals");
 const ConstDependency = require("./dependencies/ConstDependency");
+const {
+	toConstantDependency
+} = require("./javascript/JavascriptParserHelpers");
 
 /** @typedef {import("./Compiler")} Compiler */
