+++ lib/dependencies/AMDPlugin.js
@@ -5,13 +5,13 @@
 
 "use strict";
 
+const RuntimeGlobals = require("../RuntimeGlobals");
 const {
 	approve,
 	evaluateToIdentifier,
 	evaluateToString,
 	toConstantDependency
-} = require("../JavascriptParserHelpers");
-const RuntimeGlobals = require("../RuntimeGlobals");
+} = require("../javascript/JavascriptParserHelpers");
 
 const AMDDefineDependency = require("./AMDDefineDependency");
 const AMDDefineDependencyParserPlugin = require("./AMDDefineDependencyParserPlugin");
