+++ lib/dependencies/CommonJsPlugin.js
@@ -22,7 +22,7 @@ const {
 	evaluateToIdentifier,
 	evaluateToString,
 	toConstantDependency
-} = require("../JavascriptParserHelpers");
+} = require("../javascript/JavascriptParserHelpers");
 
 class CommonJsPlugin {
 	constructor(options) {
