+++ lib/dependencies/CommonJsRequireDependencyParserPlugin.js
@@ -5,8 +5,10 @@
 
 "use strict";
 
-const { toConstantDependency } = require("../JavascriptParserHelpers");
 const RuntimeGlobals = require("../RuntimeGlobals");
+const {
+	toConstantDependency
+} = require("../javascript/JavascriptParserHelpers");
 const CommonJsRequireContextDependency = require("./CommonJsRequireContextDependency");
 const CommonJsRequireDependency = require("./CommonJsRequireDependency");
 const ContextDependencyHelpers = require("./ContextDependencyHelpers");
