+++ lib/dependencies/ContextDependencyHelpers.js
@@ -5,7 +5,7 @@
 
 "use strict";
 
-/** @typedef {import("../BasicEvaluatedExpression")} BasicEvaluatedExpression */
+/** @typedef {import("../javascript/BasicEvaluatedExpression")} BasicEvaluatedExpression */
 /** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
 /** @typedef {import("./ContextDependency")} ContextDependency */
 /** @typedef {import("./ContextDependency").ContextDependencyOptions} ContextDependencyOptions */
