+++ lib/dependencies/RequireEnsurePlugin.js
@@ -13,7 +13,7 @@ const RequireEnsureDependenciesBlockParserPlugin = require("./RequireEnsureDepen
 const {
 	evaluateToString,
 	toConstantDependency
-} = require("../JavascriptParserHelpers");
+} = require("../javascript/JavascriptParserHelpers");
 
 class RequireEnsurePlugin {
 	apply(compiler) {
