+++ lib/dependencies/RequireIncludePlugin.js
@@ -11,7 +11,7 @@ const RequireIncludeDependencyParserPlugin = require("./RequireIncludeDependency
 const {
 	evaluateToString,
 	toConstantDependency
-} = require("../JavascriptParserHelpers");
+} = require("../javascript/JavascriptParserHelpers");
 
 class RequireIncludePlugin {
 	apply(compiler) {
