+++ lib/dependencies/SystemPlugin.js
@@ -5,13 +5,13 @@
 
 "use strict";
 
+const RuntimeGlobals = require("../RuntimeGlobals");
+const WebpackError = require("../WebpackError");
 const {
 	evaluateToString,
 	expressionIsUnsupported,
 	toConstantDependency
-} = require("../JavascriptParserHelpers");
-const RuntimeGlobals = require("../RuntimeGlobals");
-const WebpackError = require("../WebpackError");
+} = require("../javascript/JavascriptParserHelpers");
 const ConstDependency = require("./ConstDependency");
 const SystemRuntimeModule = require("./SystemRuntimeModule");
