+++ lib/javascript/JavascriptParser.js
@@ -8,8 +8,8 @@
 const { Parser } = require("acorn");
 const { SyncBailHook, HookMap } = require("tapable");
 const vm = require("vm");
-const BasicEvaluatedExpression = require("../BasicEvaluatedExpression");
 const StackedMap = require("../util/StackedMap");
+const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
 
 /** @typedef {import("estree").ArrayExpression} ArrayExpressionNode */
 /** @typedef {import("estree").BinaryExpression} BinaryExpressionNode */
