+++ lib/javascript/JavascriptParserHelpers.js
@@ -5,9 +5,9 @@
 
 "use strict";
 
+const UnsupportedFeatureWarning = require("../UnsupportedFeatureWarning");
+const ConstDependency = require("../dependencies/ConstDependency");
 const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
-const UnsupportedFeatureWarning = require("./UnsupportedFeatureWarning");
-const ConstDependency = require("./dependencies/ConstDependency");
 
 exports.toConstantDependency = (parser, value, runtimeRequirements) => {
 	return function constDependency(expr) {
